LayoutTests/media/video-autoplay.html

<video src=content/test.mp4 autoplay></video>
<script src=video-test.js></script>
<script>
test("video.autoplay");
test("video.getAttribute('autoplay') != null");
waitForEventTestAndEnd('play', "!video.paused");
</script>
