LayoutTests/media/video-autoplay-expected.txt

TEST(video.autoplay) OK
TEST(video.getAttribute('autoplay') != null) OK
EVENT(play) TEST(!video.paused) OK
END OF TEST
